requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pydantic-settings
pydantic[email]
pytest
pytest-asyncio
PyYAML
pypdf
pdfplumber
reportlab
pandas
openpyxl
python-docx